bench/obs: schema v2, Length stop, SIGINT-safe bench-run, first baseline, clippy gate
Changes since c03894b group into five themes:

- obs schema v2: stop.final_answer + session_start.schema_v (both
  additive, v1 traces still parse). Lets bench-replay validate
  expect.must_contain offline. Agent loop threads the last assistant
  content into both Stop emission sites.
- finish_reason="length" handling: new StopReason::Length variant.
  When llama-server reports truncation, skip dispatching potentially
  incomplete tool_calls, push a concision note into the conversation
  so the next user turn sees it, exit with Length.
- bench-run is now Ctrl-C-safe: SIGINT handler flips an atomic flag;
  children spawned with their own process group; signal handler kills
  the whole group (reaps llama-server, not just micro-mind). Timeout
  path got the same group-kill upgrade.
- First committed baseline: bench/baselines/2026-05-15-main/ (3 reps
  per task, 9 traces + summary.json). 3/9 reps pass — failures are
  real model behaviour (token-cap overrun + over-call on math).
  bench/baselines/README.md rewritten for the dir-per-run layout.
- CI now gates clippy at -D warnings (warning floor cleaned to zero
  in the process). Added an advisory step that replays every committed
  baseline against the current fixture set.

Plus a deliberate cleanup pass: ~30 collapsible if-lets folded into
let-chains, dead code removed (ChatMessage::assistant_*, three guard
reset() methods, ToolCallResult::payload, Session::new_no_recorder).

106 tests pass. cargo fmt --check and cargo clippy -D warnings both
clean. The three sample CI commands and the baseline batch replay
all produce the documented output.

+++ src/tools/shell.rs
@@ -37,7 +37,7 @@ pub fn bash(cwd: PathBuf) -> ToolDef {
                 .get("timeout_s")
                 .and_then(|v| v.as_u64())
                 .unwrap_or(30)
-                .min(300) as u64;
+                .min(300);
 
             let tokens = shlex::split(cmd_str)
                 .ok_or_else(|| format!("Could not parse command (unbalanced quote?): {cmd_str}"))?;
